Add files via upload
This assignment was locked Apr 15 at 11:59pm.
Create an HTML file with embedded JS. Put the html and images in a directory, zip it and upload the zip file. Both the initial HTML file and the ZIP file should have your name in the title. 

Rewrite the Dice example (that rolls 4 dice simultaneously) to roll 10 dice instead, using arrays. 

Define an array of 10 elements called "imgObjects" where each element is an object reference to one of the html <img> elements. 
Define an array of 6 string elements called "imgFileNames" where each element is a file name,
for example, the value in imgFileNames[0] would be "die1.png"
Rewrite function rollDice, to accept an array as the argument and use a for..in loop in the function to call setImage function for each element of the array. 
Make other necessary changes to your script.